refactor: restructure TesterApp
