+++ packages/TesterApp/src/components/asyncChunks/Async.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Text } from './Text';
+import { Text } from '../ui/Text';
 
 export default function Async() {
   return <Text>Async: this text comes from async chunk</Text>;
