+++ packages/TesterApp/src/components/miniapp/MiniAppsContainer.js
