+++ packages/TesterApp/src/components/miniapp/callstack-dark.png
