+++ packages/TesterApp/src/components/remoteChunks/Remote.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import throttle from 'lodash.throttle';
 import { StyleSheet, View } from 'react-native';
-import { Text } from './Text';
+import { Text } from '../ui/Text';
 
 export default function Remote() {
   throttle(() => {}, 0);
