+++ packages/TesterApp/src/components/ui/utils.js
@@ -1,0 +1,3 @@
+import { useColorScheme } from 'react-native';
+
+export const useIsDarkMode = () => useColorScheme() === 'dark';
